pyproject.toml
[project]
name = "healthvoicebot"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "black>=25.12.0",
    "faiss-cpu>=1.13.2",
    "flask-cors>=6.0.2",
    "flask[async]>=3.1.2",
    "livekit-agents>=1.3.10",
    "livekit-api>=1.1.0",
    "livekit-plugins-openai>=1.3.10",
    "livekit-plugins-silero>=1.3.10",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "python-frontmatter>=1.1.0",
    "sentence-transformers>=5.2.0",
    "transformers>=4.57.3",
    "uvicorn>=0.40.0",
]
